Make C-SPC work in the terminal again.
It was clashing with my keybinding for C-@, which I was using to shrink
the region. This worked fine in Windows Emacs, but not terminal Emacs,
where it stopped set-mark from working. Instead, bind shrink region to
M-'. We lose the keybinding for mark-defun, but that already exists on
the standard key C-M-h, so it is no great loss, especially considering
that I rarely use it anyway.